task 9 a-test/requirements.txt
fastapi
sqlmodel
uvicorn
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
pytest
httpx
